requirements.txt
yt-dlp
openai-whisper
google-generativeai
openai
